@@ -1,9 +1,10 @@
 {
   "name": "tailwindcss",
-  "version": "3.0.24",
+  "version": "3.1.8",
   "description": "A utility-first CSS framework for rapidly building custom user interfaces.",
   "license": "MIT",
   "main": "lib/index.js",
+  "types": "types/index.d.ts",
   "repository": "https://github.com/tailwindlabs/tailwindcss.git",
   "bugs": "https://github.com/tailwindlabs/tailwindcss/issues",
   "homepage": "https://tailwindcss.com",
@@ -25,7 +26,7 @@
     "posttest": "npm run style",
     "generate:plugin-list": "node -r @swc/register scripts/create-plugin-list.js",
     "generate:types": "node -r @swc/register scripts/generate-types.js",
-    "generate": "npm run generate:plugin-list"
+    "generate": "npm run generate:plugin-list && npm run generate:types"
   },
   "files": [
     "src/*",
@@ -35,24 +36,26 @@
     "scripts/*.js",
     "stubs/*.stub.js",
     "nesting/*",
+    "types/**/*",
+    "*.d.ts",
     "*.css",
     "*.js"
   ],
   "devDependencies": {
-    "@swc/cli": "^0.1.56",
-    "@swc/core": "^1.2.160",
-    "@swc/jest": "^0.2.20",
+    "@swc/cli": "^0.1.57",
+    "@swc/core": "^1.2.218",
+    "@swc/jest": "^0.2.22",
     "@swc/register": "^0.1.10",
-    "autoprefixer": "^10.4.4",
-    "cssnano": "^5.1.7",
-    "esbuild": "^0.14.25",
-    "eslint": "^8.11.0",
+    "autoprefixer": "^10.4.8",
+    "cssnano": "^5.1.12",
+    "esbuild": "^0.14.48",
+    "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.5.1",
-    "jest-diff": "^27.5.1",
-    "prettier": "^2.6.0",
-    "prettier-plugin-tailwindcss": "^0.1.8",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^28.1.3",
+    "jest-diff": "^28.1.3",
+    "prettier": "^2.7.1",
+    "prettier-plugin-tailwindcss": "^0.1.13",
     "rimraf": "^3.0.0",
     "source-map-js": "^1.0.2"
   },
@@ -60,27 +63,28 @@
     "postcss": "^8.0.9"
   },
   "dependencies": {
-    "arg": "^5.0.1",
+    "arg": "^5.0.2",
     "chokidar": "^3.5.3",
     "color-name": "^1.1.4",
-    "detective": "^5.2.0",
+    "detective": "^5.2.1",
     "didyoumean": "^1.2.2",
     "dlv": "^1.1.3",
     "fast-glob": "^3.2.11",
     "glob-parent": "^6.0.2",
     "is-glob": "^4.0.3",
-    "lilconfig": "^2.0.5",
+    "lilconfig": "^2.0.6",
     "normalize-path": "^3.0.0",
     "object-hash": "^3.0.0",
     "picocolors": "^1.0.0",
-    "postcss": "^8.4.12",
+    "postcss": "^8.4.14",
+    "postcss-import": "^14.1.0",
     "postcss-js": "^4.0.0",
     "postcss-load-config": "^3.1.4",
     "postcss-nested": "5.0.6",
     "postcss-selector-parser": "^6.0.10",
     "postcss-value-parser": "^4.2.0",
     "quick-lru": "^5.1.1",
-    "resolve": "^1.22.0"
+    "resolve": "^1.22.1"
   },
   "browserslist": [
     "> 1%",
@@ -96,7 +100,8 @@
     "testPathIgnorePatterns": [
       "/node_modules/",
       "/integrations/",
-      "/standalone-cli/"
+      "/standalone-cli/",
+      "\\.test\\.skip\\.js$"
     ],
     "transform": {
       "\\.js$": "@swc/jest"
