@@ -61,7 +61,9 @@ return [
     |
     */
 
-    'home' => RouteServiceProvider::HOME,
+//    'home' => RouteServiceProvider::HOME,
+    'home' => function(){return redirect("/inicio");},
+
 
     /*
     |--------------------------------------------------------------------------
@@ -134,7 +136,7 @@ return [
     'features' => [
         Features::registration(),
         Features::resetPasswords(),
-        // Features::emailVerification(),
+        Features::emailVerification(),
         Features::updateProfileInformation(),
         Features::updatePasswords(),
         Features::twoFactorAuthentication([
