@@ -5,7 +5,13 @@ export default {
 }
 </script>
 <template>
-<h1 class="text-red-400">    hola en menu</h1>
+
+  <div v-if="$page.props.auth!==null" >
+    <a v-if="$page.props.auth.roles[0]==='administrador'" class="boton" href="#">Administrar</a>
+  </div>
+  <Link class="boton" :href="route('NewUserHestia')">Crearme Usuario</Link>
+  <a class="boton" href="#">Doc Hestia</a>
+  <a class="boton" href="#">Wiki sobre hestia</a>
 </template>
 
 
