@@ -1,26 +1,27 @@
 <!DOCTYPE html>
 <html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-    <head>
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
 
-        <title inertia>{{ config('app.name', 'Laravel') }}</title>
+    <title inertia>{{ config('app.name', 'Laravel') }}</title>
 
-        <!-- Fonts -->
-        <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Nunito:wght@400;600;700&display=swap">
+    <!-- Fonts -->
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Nunito:wght@400;600;700&display=swap">
 
-        <!-- Styles -->
-        <link rel="stylesheet" href="./css/app.css">
+    <!-- Styles -->
+    <!--        <link rel="stylesheet" href="./css/app.css"  -->
+    @vite(['resources/css/app.css', 'resources/js/app.js'])
 
-        <!-- Scripts -->
-        @routes
-        <script src="./js/app.js" defer></script>
-        @inertiaHead
-    </head>
-    <body class="font-sans antialiased">
-        @inertia
-        @env ('local')
-            <script src="http://localhost:3000/browser-sync/browser-sync-client.js"></script>
-        @endenv
-    </body>
+    <!-- Scripts -->
+    @routes
+    <script src="./js/app.js" defer></script>
+    @inertiaHead
+</head>
+<body class="font-sans antialiased">
+@inertia
+@env ('local')
+    <script src="http://localhost:3000/browser-sync/browser-sync-client.js"></script>
+@endenv
+</body>
 </html>
