@@ -5,6 +5,7 @@ use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
 use App\Http\Controllers\ControladorMain;
+use App\Http\Controllers\HestiaUserController;
 
 use App\Http\Controllers\MainController;
 use App\Http\Controllers\ControladorPrueba;
@@ -23,60 +24,7 @@ use Illuminate\Http\Request;
 */
 
 
-
-
-
-Route::get('/',  [ControladorMain::class, "index"]);
-
-Route::get('/prueba1',  [ControladorMain::class, "prueba1"]);
-
-//Route::get('/',  function () {u
-//    return Inertia::render('Inicio', [
-//        'canLogin' => Route::has('login'),
-//        'canRegister' => Route::has('register'),
-//        'laravelVersion' => Application::VERSION,
-//        'phpVersion' => PHP_VERSION,
-//    ]);
-//});
-Route::get("p1", [ControladorPrueba::class,"p1"])->name("v1");
-Route::get("p2", [ControladorPrueba::class,"p2"])->name("v2");
-Route::get('/p', function () {
-    return Inertia::render('Welcome', [
-        'canLogin' => Route::has('login'),
-        'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
-    ]);
-});
-Route::inertia("login","Auth/Login")->name("login");
-
-Route::middleware([
-    'auth:sanctum',
-    config('jetstream.auth_session'),
-    'verified',
-])->group(function () {
-    Route::get('/dashboard', function () {
-        return Inertia::render('Dashboard');
-    })->name('dashboard');
-});
-Route::get('/email/verify', function () {
-    return Inertia::render('auth.verify-email');
-})->middleware('auth')->name('verification.notice');
-
-
-
-Route::get('/email/verify/{id}/{hash}', function (EmailVerificationRequest $request) {
-    $request->fulfill();
-
-    return redirect('/home');
-})->middleware(['auth', 'signed'])->name('verification.verify');
-
-
-
-
-
-Route::post('/email/verification-notification', function (Request $request) {
-    $request->user()->sendEmailVerificationNotification();
-
-    return back()->with('message', 'Verification link sent!');
-})->middleware(['auth', 'throttle:6,1'])->name('verification.send');
+Route::get('/',  [ControladorMain::class, "index"])->name("inicio");
+/*Route::get('/adentro',function(){
+    return redirect("/inicio");
+})->name("adentro");*/
